# backend-app/docs/openapi.yaml
openapi: 3.0.0
info:
  title: FlexiCharge API
  version: 1.0.0
servers:
  - url: http://localhost:8080/
    description: Development Environment
  - url: http://18.202.253.30:8080/
    description: Production Environment (AWS)
paths:
...
  /chargers/available:
    get:
      tags:
        - Chargers
      summary: Available chargers
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Charger'
        '404':
          description: Error message
  /chargers/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
    get:
      tags:
        - Chargers
      summary: Charger by id
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Charger'
        '404':
          description: Error message
        '500':
          description: Error message
    delete:
      tags:
        - Chargers
      summary: Remove charger by id
      security:
        - bearerAuth: [] # admin
      responses:
        '204':
          description: Successful response
          content:
            application/json: {}
        '404':
          description: Error message
        '500':
          description: Error message
    put:
      tags:
        - Chargers
      summary: Update charger by id
      security:
        - bearerAuth: [] # admin
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
        '400':
          description: Error message
  /chargers/serial/{serial}:
    parameters:
      - in: path
        name: serial
        schema:
          type: string
        required: true
    get:
      tags:
        - Chargers
      summary: Charger by serial number
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Charger'
        '404':
          description: Error message
  /chargers:
    post:
      tags:
        - Chargers
      summary: Chargers
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                chargerPointNumber: 23
                location:
                  - 57.777714
                  - 14.16301
                serialNumber: android
      security:
        - bearerAuth: [] # admin
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
        '404':
          description: Error message
    get:
      tags:
        - Chargers
      summary: Chargers
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Charger'
        '404':
          description: Error message
...
  /reservations/{reservationID}:
    parameters:
      - in: path
        name: reservationID
        schema:
          type: string
        required: true
    get:
      deprecated: true
      tags:
        - Reservations (deprecated)
      summary: Reservation by ID
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
    put:
      deprecated: true
      tags:
        - Reservations (deprecated)
      summary: Update Reservation
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                chargerID: 1
                connectorId: 1
                idTag: 1
                reservationID: 1
                parentIdTag: 1
      responses:
        '200':
          description: Successful response.
          content:
            application/json: {}
    delete:
      deprecated: true
      tags:
        - Reservations (deprecated)
      summary: Remove reservation
      responses:
        '204':
          description: Successful response.
          content:
            application/json: {}
        '404':
          description: Error message
        '500':
          description: Error message
  /reservations/userReservation/{userID}:
    parameters:
      - in: path
        name: userID
        schema:
          type: string
        required: true
    get:
      deprecated: true
      tags:
        - Reservations (deprecated)
      summary: Reservation by userID
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /reservations:
    post:
      deprecated: true
      tags:
        - Reservations (deprecated)
      summary: Reservation
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                chargerID: 1
                userID: 5
                start: 200
                end: 300
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /reservations/chargerReservation/{chargerID}:
    parameters:
      - in: path
        name: chargerID
        schema:
          type: string
        required: true
    get:
      deprecated: true
      tags:
        - Reservations (deprecated)
      summary: Reservation by chargerID
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...
components:
  schemas:
    Charger:
      type: object
      properties:
        chargerID:
          type: integer
          format: int64
          example: 100011
        location:
          type: array
          items:
            type: number
          example:
            - 57.777714
            - 14.16301
        serialNumber:
          type: string
          example: 'testnumber15'
        chargePointID:
          type: integer
          format: int64
          example: 24
        status:
          type: string
          description: Charger status
          example: Available
          enum:
            - Available
            - Reserved
      xml:
        name: charger
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer